requirements.txt
ipython==8.0.0
matplotlib==3.5.1
numpy==1.21.12
pandas==1.3.4
pandas-flavor==0.3.0
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
py4j==0.10.9.5
pyparsing==3.0.6
pytz==2021.3
rdkit-pypi==2021.9.4
requests==2.27.1
scikit-learn==1.0.1
scipy==1.7.3
seaborn==0.11.2
torch==1.11.0
torchaudio==0.7.2
torchvision==0.12.0
tqdm==4.62.3
urllib3==1.26.8
xarray==2022.3.0
json
